Dict/pair parsers
#4 and #2, parse into dict and pairs
The dict parser still needs a bit of work, we need to separate each value by comma first... The reason why this is on hold is because we also need to at some point check arrays and things in order to ensure that we do not split in the middle of arrays. The best way to do this is probably to use findall() to find seperators such as {} or [], and then we parse those values separately.
Started date parser.